renamed messagefrequencyvalidator class to MessageIntervalValidator
