+++ src/v2i-hub/IntersectionValidationPlugin/CMakeLists.txt
@@ -8,7 +8,7 @@ BuildTmxPlugin ()
 
 TARGET_LINK_LIBRARIES ( ${PROJECT_NAME} tmxutils ::carma-clock )
 
-add_library(${PROJECT_NAME}_lib src/MessageFrequencyValidator.cpp)
+add_library(${PROJECT_NAME}_lib src/MessageIntervalValidator.cpp)
 target_link_libraries(${PROJECT_NAME}_lib PUBLIC tmxutils)
 target_include_directories(${PROJECT_NAME}_lib PUBLIC ${PROJECT_SOURCE_DIR}/src)
